it doesn't draw

+++ src/APIs/wrappers/gl_wrapper.cpp
@@ -24,7 +24,7 @@ namespace RandysEngine{
 
         // glfw window creation
         // --------------------
-        window = glfwCreateWindow((int)width*4, (int)height*4, "Randys Engine Example", NULL, NULL);
+        window = glfwCreateWindow(width*4, height*4, "Randys Engine Example", NULL, NULL);
         if (window == NULL){
             std::cout << "Failed to create GLFW window" << std::endl;
             glfwTerminate();
